Refactor SmartSim implementation
The SmartSim implementation was being run as a python script. For
future flexibility and scaling, this has now been refactored to
be a "true" SmartSim-like experiment. The inferencers can now be
replicated to mimic multiple MPI ranks calling for inference and
batching across requests can be optionally enabled.

Additionally, the training and benchmarking scripts have been
been updated to allow for GPU-accelerated training and inference.